issue 877 - moved request authenticator & factory back to http/authc
